reporter: Add scan summary to PackagesTable
